requirements.txt
openai>=1.30.0
llama-index>=0.11.0
llama-index-llms-openai>=0.2.0
llama-index-embeddings-openai>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
rich>=13.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
